Decision Trees
A decision tree can be used for both classification and regression problems.
A decision tree is a flowchart-like tree structure where an internal node represents feature(or attribute),
the branch represents a decision rule, and each leaf node represents the outcome. 
The dataset that was used consists of several medical predictor variables and one target variable, Outcome. Predictor variables includes the number of
pregnancies the patient has had, their BMI, insulin level, age, and so on