Hierarchical Clustering
Hierarchical clustering is an unsupervised clustering method that you can use to group your data. This
algorithm is unsupervised because it uses a random, unlabelled dataset. The resulting clusters are displayed
in a hierarchical tree called a dendrogram. This is helpful because the algorithm produces a clear graphical
depiction of your clusters that you can understand and interpret easily. With this algorithm, you can create
decision trees as well as category hierarchies for your business.