Logistics Regression
Logistic regression is a supervised classification algorithm. Although it can be extended to classify among
more than 2 classes, the algorithm in its nature can only predict if a data point belongs to a class or not.
Logistic regression is a statistical method for predicting binary classes. The outcome or target variable is
dichotomous in nature. Dichotomous means there are only two possible classes. For example, it can be used for
diabetes detection problems. It computes the probability of an event occurrence.