K Means Clustering
K-Means clustering is an unsupervised algorithm used to divide the data into `K` number of clusters based on
the distance of the each point from the mean (could be arithmetric mean) from the centroid of the cluster. In
other words, the algorithm uses the distance between points as a measure of similarity, based on k averages
(i.e. means)